requirements.txt
azure-functions
scikit-learn
pandas
joblib
numpy
scipy
